:truck: Move I/O helpers under util
